auth added.

+++ src/main/java/tr/unvercanunlu/todoapp/controller/ToDoController.java
@@ -1,9 +1,14 @@
 package tr.unvercanunlu.todoapp.controller;
 
+import static tr.unvercanunlu.todoapp.config.AuthConfig.ROLE_HEADER;
+import static tr.unvercanunlu.todoapp.config.AuthConfig.ROLE_USER;
+import static tr.unvercanunlu.todoapp.config.AuthConfig.TOKEN_HEADER;
+
 import java.net.URI;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -11,11 +16,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import tr.unvercanunlu.todoapp.model.ToDo;
 import tr.unvercanunlu.todoapp.model.ToDoRequest;
 import tr.unvercanunlu.todoapp.service.ToDoService;
+import tr.unvercanunlu.todoapp.util.ValidationUtil;
 
 @RestController
 @RequestMapping("/todos")
@@ -23,6 +30,7 @@ import tr.unvercanunlu.todoapp.service.ToDoService;
 public class ToDoController {
 
   private final ToDoService toDoService;
+  private final ValidationUtil validationUtil;
 
   @PostMapping
   public ResponseEntity<ToDo> createToDo(@RequestBody ToDoRequest request) {
@@ -44,7 +52,16 @@ public class ToDoController {
   }
 
   @PutMapping("/{id}")
-  public ResponseEntity<ToDo> updateToDo(@PathVariable Long id, @RequestBody ToDoRequest request) {
+  public ResponseEntity<ToDo> updateToDo(@PathVariable Long id, @RequestBody ToDoRequest request,
+      @RequestHeader(name = TOKEN_HEADER, required = false) String token,
+      @CookieValue(name = ROLE_HEADER, defaultValue = ROLE_USER, required = false) String role) {
+
+    // HTTP request header / HTTP cookie based authorization control
+    if (!(validationUtil.isTokenMatched(token) || validationUtil.isRoleMatched(role))) {
+      return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+          .build();
+    }
+
     ToDo updated = toDoService.updateToDo(id, request);
 
     return ResponseEntity.status(HttpStatus.OK)
@@ -68,7 +85,16 @@ public class ToDoController {
   }
 
   @DeleteMapping("/{id}")
-  public ResponseEntity<Void> deleteToDo(@PathVariable Long id) {
+  public ResponseEntity<Void> deleteToDo(@PathVariable Long id,
+      @RequestHeader(name = TOKEN_HEADER, required = false) String token,
+      @CookieValue(name = ROLE_HEADER, defaultValue = ROLE_USER, required = false) String role) {
+
+    // HTTP request header / HTTP cookie based authorization control
+    if (!(validationUtil.isTokenMatched(token) || validationUtil.isRoleMatched(role))) {
+      return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+          .build();
+    }
+
     toDoService.deleteToDo(id);
 
     return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
